FEAT(Three-id): Extract Profile Details screen
F&L
